Use a slightly more common test framework
This brings FR in line with the rest of Twilio's projects in terms of
installing and running the tests. Please let me know if this causes issues.

You can now run the tests with:

    make clean install test-install test

which will do a clean install and test. The next time you run the tests you can
simply use "make test" to do the job.