Refactor tests to remove repetition.
This is a superficial refactor of the tests to be a bit more streamlined
by removing obvious repetition from the test classes. Most of this is
done by simply adding a Flask app instance with an API in the setUp
method of the TestCase classes. The tests in terms of what they do have
been unchanged.

This provides slightly simpler code to read - and provides new tests
with pre setup app and api instances if needed.

Also a minor refactor of import orderings in tests to match PEP8.