Refactor tests to reduce repetition and use py.test over nose.
Nose has been deprecated for several years whereas py.test is still
under active development. As a result it is more accepted nowadays to
use py.test as the test runner.

This refactor also reduces much of the repetition in the tests and moves
to the functional style of testing favoured by py.test.