Stop passing the field name to the Decimal constructor
When special-casing the request parser for Decimal, passing the name of
the field as second argument is incorrect.

The second argument of Decimal is a context. Python 3 checks this is in
fact a context, whereas Python 2 ignores it if not needed. This means in
Python 3 that special case line is never successful, so it is handled
below. However, there is a difference: the first converts to string, the
second does not. This results in different values of Decimal being
parsed.